% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daybreak-wrappers.R
\name{astronomical_twilight}
\alias{astronomical_twilight}
\title{Astronomical twilight}
\usage{
astronomical_twilight(date, lon, lat)
}
\arguments{
\item{date}{The date to compute the length for. An R \link{DateTimeClasses} object
or something that can be coerced into one by \code{\link[=as.POSIXlt]{as.POSIXlt()}}.}

\item{lon, lat}{longitude & latitude}
}
\value{
(dbl) astronomical twilight
}
\description{
Astronomical twilight
}
\examples{
astronomical_twilight("2019-12-31", -70.8636, 43.2683)
}
