% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daybreak-wrappers.R
\name{day_astronomical_twilight_length}
\alias{day_astronomical_twilight_length}
\title{Length of astronomical twilight}
\usage{
day_astronomical_twilight_length(date, lon, lat)
}
\arguments{
\item{date}{The date to compute the length for. An R \link{DateTimeClasses} object
or something that can be coerced into one by \code{\link[=as.POSIXlt]{as.POSIXlt()}}.}

\item{lon, lat}{longitude & latitude}
}
\value{
(dbl) astronomical twilight length
}
\description{
Length of astronomical twilight
}
\examples{
day_astronomical_twilight_length("2019-12-31", -70.8636, 43.2683)
}
