% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daybreak-wrappers.R
\name{day_length}
\alias{day_length}
\title{Length of day}
\usage{
day_length(date = Sys.Date(), lon, lat)
}
\arguments{
\item{date}{The date to compute the length for. An R \link{DateTimeClasses} object
or something that can be coerced into one by \code{\link[=as.POSIXlt]{as.POSIXlt()}}.}

\item{lon, lat}{longitude & latitude}
}
\value{
(dbl) length of day
}
\description{
Length of day
}
\examples{
day_length("2019-12-31", -70.8636, 43.2683)
}
