% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daybreak-package.R
\docType{package}
\name{daybreak}
\alias{daybreak}
\alias{daybreak-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/daybreak}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/daybreak/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
