% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daybreak-wrappers.R
\name{nautical_twilight}
\alias{nautical_twilight}
\title{Nautical twilight}
\usage{
nautical_twilight(date, lon, lat)
}
\arguments{
\item{date}{The date to compute the length for. An R \link{DateTimeClasses} object
or something that can be coerced into one by \code{\link[=as.POSIXlt]{as.POSIXlt()}}.}

\item{lon, lat}{longitude & latitude}
}
\value{
(dbl) nautical twilight
}
\description{
Nautical twilight
}
\examples{
nautical_twilight("2019-12-31", -70.8636, 43.2683)
}
