% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daybreak-wrappers.R
\name{sun_rise_set}
\alias{sun_rise_set}
\title{Sun rise/set times}
\usage{
sun_rise_set(date, lon, lat)
}
\arguments{
\item{date}{The date to compute the length for. An R \link{DateTimeClasses} object
or something that can be coerced into one by \code{\link[=as.POSIXlt]{as.POSIXlt()}}.}

\item{lon, lat}{longitude & latitude}
}
\value{
(dbl) sunrise/sunset
}
\description{
Sun rise/set times
}
\examples{
sun_rise_set("2019-12-31", -70.8636, 43.2683)
}
