% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daybreak-package.R
\docType{package}
\name{daybreak}
\alias{daybreak}
\alias{daybreak-package}
\title{Compute Sun Rise/Set Times, Start/End of Twilight, and the
Length of the Day at Any Date and Latitude}
\description{
A wrapper for Paul Schlyter's C-based library for computing
sunrise, sunset, twilight start and end, plus the length of day for
a given data and coordinates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/daybreak}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/daybreak/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
